Migrate ValueValidators to BigInteger

+++ src/main/java/abstractalgebra/ints/AllIntsValidator.java
@@ -1,5 +1,7 @@
 package abstractalgebra.ints;
 
-public class AllIntsValidator implements ValueValidator<Integer> {
-    @Override public void assertValid(Integer v) { }
+import java.math.BigInteger;
+
+public class AllIntsValidator implements ValueValidator<BigInteger> {
+    @Override public void assertValid(BigInteger v) { }
 }
